Remove cyclic import so test pass on 3.6
Interestingly in the move to python 3.7 they patched some cyclic
import issues.
