+++ src/openfermion/linalg/__init__.py
@@ -32,8 +32,7 @@ from .sparse_tools import (
     jw_hartree_fock_state, jw_number_indices, jw_sz_indices,
     jw_number_restrict_operator, jw_sz_restrict_operator,
     jw_number_restrict_state, jw_sz_restrict_state,
-    jw_get_ground_state_at_particle_number, jw_get_gaussian_state,
-    jw_slater_determinant, jw_sparse_givens_rotation,
+    jw_get_ground_state_at_particle_number, jw_sparse_givens_rotation,
     jw_sparse_particle_hole_transformation_last_mode, get_density_matrix,
     get_ground_state, sparse_eigenspectrum, expectation, variance,
     expectation_computational_basis_state,
