+++ src/openfermion/utils/commutators.py
@@ -14,7 +14,7 @@
 import numpy
 
 from openfermion.ops.operators import FermionOperator
-import openfermion.utils.operator_utils as op_utils
+from openfermion.utils.operator_utils import normal_ordered
 
 
 def commutator(operator_a, operator_b):
@@ -105,9 +105,9 @@ def double_commutator(op1,
             ((index2, 1), (index3, 0)), coeff2 * coeff3) + FermionOperator(
                 ((index3, 1), (index2, 0)), -coeff2 * coeff3))
     else:
-        commutator23 = op_utils.normal_ordered(commutator(op2, op3))
+        commutator23 = normal_ordered(commutator(op2, op3))
 
-    return op_utils.normal_ordered(commutator(op1, commutator23))
+    return normal_ordered(commutator(op1, commutator23))
 
 
 def trivially_double_commutes_dual_basis_using_term_info(
